Baseline/requirements.txt
torch
h5py
numpy
Pillow
matplotlib
torchvision
googledrivedownloader